processors/spacyentities/requirements.txt
spacy
nats-py
